Revert #140650 and #140505
This reverts commit 90daed32a82ad2695d27db285ac36f579f2b270e and
4cfbe55781cb8fb95568c9a8538912f68d2ff681.
These changes exposed cyclic dependencies when LLVM is configured with
modules `-DLLVM_ENABLE_MODULES=ON`.
